requirements.txt
aiogram==3.3.0
gigachat==0.1.3
python-docx==1.1.0
python-dotenv==1.0.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
soundfile>=0.12.0
accelerate>=0.25.0  
ffmpeg-python>=0.2.0  